+++ bin/domain/usecases/HandleScheduledEventUseCase.js
@@ -9,7 +9,7 @@ class ProjectNotFoundError extends Error {
 }
 exports.ProjectNotFoundError = ProjectNotFoundError;
 class HandleScheduledEventUseCase {
-    constructor(generateWorkingTimeReportUseCase, actionAnnouncementUseCase, setWorkflowManagementIssueToStoryUseCase, clearNextActionHourUseCase, analyzeProblemByIssueUseCase, analyzeStoriesUseCase, clearDependedIssueURLUseCase, createEstimationIssueUseCase, convertCheckboxToIssueInStoryIssueUseCase, changeStatusLongInReviewIssueUseCase, changeStatusByStoryColorUseCase, setNoStoryIssueToStoryUseCase, createNewStoryByLabelUseCase, dateRepository, spreadsheetRepository, projectRepository, issueRepository) {
+    constructor(generateWorkingTimeReportUseCase, actionAnnouncementUseCase, setWorkflowManagementIssueToStoryUseCase, clearNextActionHourUseCase, analyzeProblemByIssueUseCase, analyzeStoriesUseCase, clearDependedIssueURLUseCase, createEstimationIssueUseCase, convertCheckboxToIssueInStoryIssueUseCase, changeStatusLongInReviewIssueUseCase, changeStatusByStoryColorUseCase, setNoStoryIssueToStoryUseCase, createNewStoryByLabelUseCase, assignNoAssigneeIssueToManagerUseCase, dateRepository, spreadsheetRepository, projectRepository, issueRepository) {
         this.generateWorkingTimeReportUseCase = generateWorkingTimeReportUseCase;
         this.actionAnnouncementUseCase = actionAnnouncementUseCase;
         this.setWorkflowManagementIssueToStoryUseCase = setWorkflowManagementIssueToStoryUseCase;
@@ -23,6 +23,7 @@ class HandleScheduledEventUseCase {
         this.changeStatusByStoryColorUseCase = changeStatusByStoryColorUseCase;
         this.setNoStoryIssueToStoryUseCase = setNoStoryIssueToStoryUseCase;
         this.createNewStoryByLabelUseCase = createNewStoryByLabelUseCase;
+        this.assignNoAssigneeIssueToManagerUseCase = assignNoAssigneeIssueToManagerUseCase;
         this.dateRepository = dateRepository;
         this.spreadsheetRepository = spreadsheetRepository;
         this.projectRepository = projectRepository;
@@ -218,6 +219,11 @@ ${JSON.stringify(e)}
                 repo: input.workingReport.repo,
                 disabledStatus: input.disabledStatus,
                 storyObjectMap: storyObjectMap,
+            });
+            await this.assignNoAssigneeIssueToManagerUseCase.run({
+                issues,
+                manager: input.manager,
+                cacheUsed,
             });
         };
         this.runForGenerateWorkingTimeReportUseCase = async (input) => {