+++ types/adapter/repositories/GraphqlProjectRepository.d.ts
@@ -1,7 +1,7 @@
 import { BaseGitHubRepository } from './BaseGitHubRepository';
 import { ProjectRepository } from '../../domain/usecases/adapter-interfaces/ProjectRepository';
-import { FieldOption, Project } from '../../domain/entities/Project';
-export declare class GraphqlProjectRepository extends BaseGitHubRepository implements ProjectRepository {
+import { Project } from '../../domain/entities/Project';
+export declare class GraphqlProjectRepository extends BaseGitHubRepository implements Pick<ProjectRepository, 'getProject' | 'findProjectIdByUrl'> {
     extractProjectFromUrl: (projectUrl: string) => {
         owner: string;
         projectNumber: number;
@@ -9,8 +9,5 @@ export declare class GraphqlProjectRepository extends BaseGitHubRepository imple
     fetchProjectId: (login: string, projectNumber: number) => Promise<string>;
     findProjectIdByUrl: (projectUrl: string) => Promise<Project["id"] | null>;
     getProject: (projectId: Project["id"]) => Promise<Project | null>;
-    addNewStory: (project: Project, storyOptions: (Omit<FieldOption, "id"> & {
-        id: FieldOption["id"] | null;
-    })[]) => Promise<FieldOption[]>;
 }
 //# sourceMappingURL=GraphqlProjectRepository.d.ts.map