+++ types/domain/usecases/AssignNoAssigneeIssueToManagerUseCase.d.ts
@@ -1,13 +1,12 @@
 import { Issue } from '../entities/Issue';
 import { IssueRepository } from './adapter-interfaces/IssueRepository';
-import { Project } from '../entities/Project';
+import { Member } from '../entities/Member';
 export declare class AssignNoAssigneeIssueToManagerUseCase {
-    readonly issueRepository: Pick<IssueRepository, 'updateStory'>;
-    constructor(issueRepository: Pick<IssueRepository, 'updateStory'>);
+    readonly issueRepository: Pick<IssueRepository, 'updateAssigneeList'>;
+    constructor(issueRepository: Pick<IssueRepository, 'updateAssigneeList'>);
     run: (input: {
-        targetDates: Date[];
-        project: Project;
         issues: Issue[];
+        manager: Member["name"];
         cacheUsed: boolean;
     }) => Promise<void>;
 }